Create Bkp_6th_dec_Best_Prod_code_latest_index.html
This code is BKP code of the current deployed code from prod. I have make some changes in the next file from this code where we can fetch all the OAID & details with the help of Override Id & MFL Id from the API stream. It will retrieve the whole row matching the combination.